Removed storyboard dependency for iconView Image.
The previous code required the user to create a UIImageView IBOutlet
and connect this outlet to the iconView property.

Like all IBOutlets, the IBOutlet property in the VPRubberCell class can
only be connected to a SINGLE image view at any time; because of this
only one controller in storyboard could use the VPRubberMenu with the
iconView.

I have removed this dependency.